Configure alt/cmd swap for macbook keyboard
By default, Macbook keyboard maps "cmd" button to "meta" and "alt" to
"alt". It seems sane but cmd button on Macbooks is where alt button
should be [1] and alt is located where you expect to find the "meta"
button. So it's not convenient for me.

This commit adds swapping of alt/cmd buttons for macbook. We have to do
this via `when` condition because gsettings can't append value to the
key - it will overwrite it.

[1]: https://www.parallels.com/blogs/app/uploads/2018/01/Blog_ALT_1078x516_2018.png