terminal: Remove configuration via profile settings
Previously we were configuring terminal by setting options in default
profile via dconf. But it turns out that profile settings are not
applied! These writes via dconf doesn't have any effect. Note, though,
that settings changed via GUI menu are reflected in dconf :shrug:.
So it seems that dconf configuration for profile is one way - changes
are reflected in dconf but not applied when you edit dconf.
It may be like that because in dconf all profile settings are shown with
"No Schema Found" comment.

This commit removes dconf configuration for profile settings of the
terminal because they are not applied via dconf.