Add ramfs rebuild for hid_apple kernel module
When we configure hid_apple kernel module for Macbook keyboard settings,
we apply settings immediately by doing rmmod and modprobe. But the
settings are not persisted on reboot. The reason is that hid_apple is
part of initramfs, therefore we need to rebuild initramfs with hid_apple
with updated params.

In Fedora initramfs is managed by dracut utility and requires -f option
to overwrite current initramfs.

This commit adds handler for rebuilding initramfs via dracut on
hid_apple configuration changes.