dotfiles: Add AppArmor override for redshift on Ubuntu
On Ubuntu redshift package installs AppArmor profile that restrict
access to everything except ~/.config/redshift.conf and some dbus calls.
I deploy this config as symlink to a file in dotfiles repo
(~/dev/dotfiles) and AppArmor blocks the access to it. Since I don't
want to install redshift in the other way, the fix is to override
AppArmor profile. The override allows access just to redshift.conf in
dotfiles repo.